# api/openapi.yaml
openapi: 3.0.3
info:
  title: Books API
  description: A production-ready RESTful API for managing books with PostgreSQL and Redis
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/yourusername/load-test
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: books
    description: Book management operations

paths:
  /books:
    get:
      summary: Get all books
      description: Retrieve all books (cached for 5 minutes in Redis)
      operationId: getAllBooks
      tags:
        - books
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new book
      description: Create a new book with title, author, and ISBN
      operationId: createBook
      tags:
        - books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /books/{id}:
    get:
      summary: Get a book by ID
      description: Retrieve a single book by its ID (cached for 5 minutes in Redis)
      operationId: getBookByID
      tags:
        - books
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update a book
      description: Update an existing book by ID
      operationId: updateBook
      tags:
        - books
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete a book
      description: Delete a book by ID
      operationId: deleteBook
      tags:
        - books
      parameters:
        - name: id
          in: path
          required: true
          description: Book ID
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Book deleted successfully
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - author
        - isbn
      properties:
        id:
          type: integer
          format: int32
          description: Unique book identifier
          example: 1
        title:
          type: string
          description: Book title
          example: "The Go Programming Language"
          minLength: 1
          maxLength: 255
        author:
          type: string
          description: Book author
          example: "Alan Donovan"
          minLength: 1
          maxLength: 255
        isbn:
          type: string
          description: ISBN number
          example: "978-0134190440"
          pattern: '^[0-9-]+$'
          minLength: 10
          maxLength: 20
    
    BookInput:
      type: object
      required:
        - title
        - author
        - isbn
      properties:
        title:
          type: string
          description: Book title
          example: "The Go Programming Language"
          minLength: 1
          maxLength: 255
        author:
          type: string
          description: Book author
          example: "Alan Donovan"
          minLength: 1
          maxLength: 255
        isbn:
          type: string
          description: ISBN number
          example: "978-0134190440"
          pattern: '^[0-9-]+$'
          minLength: 10
          maxLength: 20
    
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
          example: "Book not found"