# internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package database

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Book struct {
	ID        int32              `json:"id"`
	Title     string             `json:"title"`
	Author    string             `json:"author"`
	Isbn      string             `json:"isbn"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}